feat: Add build script for Schedule_Genius
The build.sh script is a new addition to the repository, providing a convenient way to build and run the Schedule_Genius application. The script uses `cmake` to build the project in the `build` directory, then changes to the `build` directory and executes the `Schedule_Genius` executable. Finally, it returns to the parent directory.