feat: Modify build script and TabDynChar class implementation
This commit modifies the build script by changing the directory to 'build/' before executing the 'Schedule_Genius' command. It also updates the TabDynChar class by adding a new member variable 'm_tab' and initializing it with an empty array of characters. Additionally, it modifies the copy constructor of the TabDynChar class to use the 'Add' method instead of directly assigning values to the 'm_tab' array.