feat: Update dictionary classes
This commit updates the dictionary classes `DictDynCharDouble`, `DictDynIntUnsignedFloat`, `DictDynCharInt`, and `DictDynIntUnsignedEleve`. The changes involve adding new functionalities and fixing some issues with value and key comparison. The modifications include the addition of new parameters in the `IfValue()` and `IfKey()` functions, as well as the implementation of the `TestIfValue()` and `TestIfKey()` functions. This commit also includes some minor formatting changes.